Refactor form definition schemas to remove cycles
