+++ src/core/findInputDefinitionSchema.ts
@@ -1,8 +1,9 @@
 import { A, NonEmptyArray, ParsingFn, parse, pipe } from "@std";
 import * as E from "fp-ts/Either";
 import { BaseSchema, ValiError } from "valibot";
-import { AllFieldTypes } from "./formDefinition";
-import { FieldMinimal, FieldMinimalSchema } from "./formDefinitionSchema";
+import type { FieldMinimal } from "./fieldDefinitionSchema";
+import { FieldMinimalSchema } from "./fieldDefinitionSchema";
+import type { AllFieldTypes } from "./input/inputDefinitionTypes";
 import { InputTypeToParserMap } from "./input/InputDefinitionSchema";
 
 export function stringifyIssues(error: ValiError): NonEmptyArray<string> {
