+++ src/core/input/inputDefinitionTypes.ts
@@ -1,0 +1,23 @@
+export type AllFieldTypes =
+    | "text"
+    | "number"
+    | "date"
+    | "time"
+    | "datetime"
+    | "textarea"
+    | "toggle"
+    | "email"
+    | "tel"
+    | "select"
+    | "tag"
+    | "note"
+    | "folder"
+    | "slider"
+    | "dataview"
+    | "multiselect"
+    | "document_block"
+    | "markdown_block"
+    | "image"
+    | "file";
+
+export type AllSources = "notes" | "fixed" | "dataview";
